requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.13
langchain-core>=0.1.10
openai>=1.0.0
faiss-cpu>=1.7.4
numpy>=1.24.0
pandas>=2.0.0
owlready2>=0.46
neo4j>=5.0.0
streamlit>=1.31.0
plotly>=5.18.0
koreanize-matplotlib>=0.1.2
anthropic>=0.18.0
python-dotenv>=1.0.0
